Implement another way to parse OSSL_PARAM keys
The idea is to pre-generate a parse-tree for all params in a single
provider.  This implementation creates a TRIE from all the param keys,
which are specified in a file, and implements it as a single function
(TRIE-as-code) that returns a parameter key index, along with a set of
macros that give names to those indexes.