cmake/provider.cmake: Give up on trying to fix FindOpenSSL.cmake's mess
A lot of effort went into trying to fix FindOpenSSL.cmake's mixup of custom
and Shining Light OpenSSL installations on Windows.  Unfortunately, that
ended up creating more of a mess.  It's therefore time to admit defeat.

The 'setup_provider_openssl' macro still has a purpose, though, to fill in
the missing parts, such as library and modules directories, the path of the
openssl program, etc.

With MSVC, if you know that the Shining Light distribution is installed, and
you want to try to use a custom install, simply do not build with the Debug
configuration, because that's what FindOpenSSL.com doesn't get right.  All
other configurations should use the custom installation's libraries.

Note: the information given here relies on the library target property
IMPORTED_LOCATION_DEBUG only being used for the Debug configuration,
while the property IMPORTED_LOCATION_RELEASE is used for all other
configurations.