cmake/provider.cmake: add more variables
For providers that link with a shared libcrypto on Windows, it's necessary
to copy the OpenSSL DLLs to the directory where the provider resides, 'cause
that's how Windows works.  To that end, this macro now also provides the
following variables if they aren't already defined:

OPENSSL_RUNTIME_DIR             The runtime directory, i.e. where openssl.exe
                                resides
OPENSSL_LIBCRYPTO_SHARED        The libcrypto DLL file name
OPENSSL_LIBSSL_SHARED           The libssl DLL file name

This is not provided for non-Windows platforms, as it's not necessary there.